base.d.js export trick for types
